// src/app/models/registros.ts
export interface Registro {
    id?: number;
    fechareg?: Date;
    servicio?: string;
    tiposervicio?: string;
    nit?: string;
    nombreemp?: string;
    teleemp?: string;
    tipodoc?: string;
    primernom?: string;
    segundonom?: string;
    primerape?: string;
    segundoape?: string;
    ciudad?: string;
    munlocalidad?: string;
    fechanac?: string;
    telcel?: string;
    correo?: string;
    observacion?: string;
    usuario?: string;
    origen?: string;
}